camera: do not try to recompute the up vector
This chunk of code has never worked as intended as it computes the up
vector from the initial eye, center and ground coordinates without
taking into account the potential transformations.
Moreover, even if we recompute the up vector with the eye and center
transformations applied we still can end up in a scenario where the up
vector is parallel to the line of sight.

It is up to the user to provide a proper up vector depending on the
other parameters.

This partially reverts e67382ad86f60627ca4e3cad52510679565911e5.

(cherry picked from commit 654835ba7dfa6ca4343ce4cbe5c145e332ff69e8)