build/nodegl: bugfix: don't override search directory when calling cc.find_library.
By default, cc.find_library searches for a library in the system library paths.
When passing dirs param, it will check in the user-specified directories instead.
With this fix, meson first searches for the library in the system path, then
if it can't find the library it will search in the user-specified path.

Resolves issue:
Library OpenGL32 found: NO
Library gdi32 found: NO