ngfx-integ/gctx_ngfx: update logic to handle when capture_buffer is NULL
When capture_buffer is NULL, we still need to issue the commands.
Resolves validation error in Vulkan backend:
expects VkImage to be in layout ..., instead current layout is ... .
When we record a layout transition, we store the new image layout as a
state variable.  If the command buffer is never submitted, then the
command is never executed, and the state that we store is incorrect,
and the Vulkan validation layer will complain.