api: only start the default renderpass when needed
The graphics context now provides two default rendertargets. The first
default rendertarget is configured with load operations set to
LOAD_OP_CLEAR and store operations set to LOAD_OP_STORE while the second
one is configured with load operations set to LOAD_OP_LOAD and store
operations set to STORE_OP_STORE.

This allows us to remove the explicit clear command in gl_begin_draw()
and starts the main renderpass only when it is needed.

This improves the following scenario:

                  .---(Group)---.
                  |             |
                  |             |
                  |             |
                (RTT 1)     (Render 1)

Before this commit:
- the main renderpass (LOAD_OP=LOAD, STORE_OP=STORE) starts and an
  explicit clear command is executed in gl_begin_draw()
- the main renderpass ends and the renderpass associated with RTT 1 is
  executed
- the main renderpass (LOAD_OP=LOAD, STORE_OP=STORE) restarts and the
  pipeline associated with Render 1 is executed
- the main renderpass ends and its color attachment is presented to
  screen

With this commit:
- the renderpass associated with RTT 1 is executed
- the main renderpass (LOAD_OP=CLEAR, STORE_OP=STORE) starts, the
  pipeline associated with Render 1 is executed
- the main renderpass ends and its color attachment is presented to
  screen

NOTE: the renderpass depth/stencil store operations are not optimal
because it would require to analyze the graph to infer whether we can use
NGLI_STORE_OP_DONT_CARE or not and setup the renderpass accordingly.