glstate: reset the VAO state
It is useful to reset the VAO state if the OpenGL context is external
(wrapped) as we have no control on how it was created and what external
commands are made. In particular, the external context could be created
using a compatibility profile which do not require the use of VAOs
meaning external commands (made outside of node.gl) can affect the
vertex attributes global state. This is an issue as node.gl does not
reset the VAO state between draw calls leaving the last used VAO bound,
meaning that any external commands affecting the vertex attributes state
will clobber the state of the last VAO bound by node.gl. This commit
prevents this from happening by resetting the VAO state while resetting
the overall OpenGL state in ngli_glstate_reset().