texture: add NGLI_TEXTURE_USAGE_TRANSIENT_ATTACHMENT_BIT
Using NGLI_TEXTURE_USAGE_TRANSIENT_ATTACHMENT_BIT specifies that
implementations (that support it) may use special allocations (lazy
allocations) to back a texture. The content of such lazily allocated
textures is only available/valid for the duration of the render pass the
textures are attached to. Using
NGLI_TEXTURE_USAGE_TRANSIENT_ATTACHMENT_BIT for any attachment that is
not re-used helps tile-based GPU to save bandwidth.

NGLI_TEXTURE_USAGE_TRANSIENT_ATTACHMENT_BIT is currently a no-op on the
OpenGL backend but it will potentially unlock the use of
EXT_multisampled_render_to_texture and
EXT_multisampled_render_to_texture2 in the future.