nodes: remove UniformColorA
The semantic of the alpha channel is unclear with regard to the color
(is it premultiplied or unassociated alpha?).

The colors are now always expressed as triplet coordinates in node
parameters, so having a 4-component long color is not very useful and
can be replaced with a non-ambiguous alternative: a separate opacity
channel, which will need to be interpolated differently anyway.

The removal of this node will simplify future commit. In particular, it
will prevent the need to add a ton of "A" variants of the same nodes.