pass: add support for resources drawing
We want to add support for ColorStats in Render and Compute, but
ColorStats needs its draw() callback to be called before its data can be
read.

To address this, we look into each resource and check whether they have
a draw callback, in which case the resource is registered and its draw
callback executed in ngli_pass_exec().