node_renderother: fix for D3D12 issue: Vertex Shader - Pixel Shader linkage error: Signatures between stages are incompatible.
The problem is that output signature of the vertex shader doesn't
match the input signature of the pixel shader.

Here's the output signature of the vertex shader:

; Name                 Index   Mask Register SysValue  Format   Used
; -------------------- ----- ------ -------- -------- ------- ------
; TEXCOORD                 0   xy          0     NONE   float   xy
; TEXCOORD                 1     zw        0     NONE   float     zw
; SV_Position              0   xyzw        1      POS   float   xyzw

Here's the input signature of the fragment shader:

; Name                 Index   Mask Register SysValue  Format   Used
; -------------------- ----- ------ -------- -------- ------- ------
; TEXCOORD                 1   xy          0     NONE   float   xy
;

TEXCOORD 1 has register 0, mask zw in vertex shader output, and register
0 mask xy is pixel shader input.  The mask doesn't match.  The root
cause is that the vertex shader outputs data that isn't used by the
pixel shader.  The solution is to either strip the vertex shader so
it doesn't output unnecessary data, or update the vertex shader output
struct and pixel shader input struct so that the pixel shader input
struct is a subset of the vertex shader output struct.