tests/compute: don't use keyword gl_NumWorkGroups.
HLSL does not have an equivalent keyword to query the number of
workgroups.  An alternative approach is to call the spirv-cross function
remap_num_workgroups_builtin but that is a more complicated approach:
it adds a dummy buffer with the number of workgroups but it also
requires the calling application to create and pass this buffer.