build: add support for cross-compilation on macOS
macOS now supports cross-compiling for arm64 and x86_64 architectures.
But we get the following error:
> Tried to install a target for the build machine in a cross build.

The problem is that the flag `is_native` wasn't correctly set.

The proposed solution is to rely on `meson.is_cross_build()` to set the
`is_native` flag, instead of manually setting it depending on the
`host_system`.