Build: Transition leak and release scripts from sh to python.
In order to make the utility scripts of the project more portable, sh scripts are being replaced with python scripts. The replacement scripts were introduced some time ago and have been in use and tested. This commit makes the transition official by updating documentation and removing the old sh scripts.